+++ docs/source/changelog.rst
@@ -1577,7 +1577,10 @@ Changes
   - The NLP cloud client checks that all requested processors are available
     remotely, and fails overtly, rather than silently ignoring those.
 
-  - Alcohol units regex NLP.
+  - Alcohol units-per-week regex NLP.
+
+  - Support "over" and "under" as further synonyms for inequalities in regex
+    NLP.
 
 - Bayesian linkage, identifiable and de-identified.